infra/docker/scraper/requirements.txt
fastapi
uvicorn[standard]
httpx
newspaper3k
lxml[html_clean]>=5.2.0